Fixes validators
